check auth on header
